Rename errors::http to errors::json
The remaining helper structs that build json errors were also moved to
this module.
